components/tableIndexer/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorlog==6.8.2
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
easyocr==1.7.1
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==23.2
pdf2image==1.17.0
pillow==10.2.0
pyclipper==1.3.0.post5
pyparsing==3.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sympy==1.12
tifffile==2024.2.12
tokenizers==0.15.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.0
yarg==0.1.9
